backend-thea-aa/requirements.txt
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.1
Django==5.1.2
django-cors-headers==4.7.0
django-crontab==0.7.1
django-filter==24.3
djangorestframework==3.15.2
dotenv==0.9.9
h11==0.14.0
idna==3.10
Markdown==3.7
outcome==1.3.0.post0
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.32.3
selenium==4.25.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.1
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.3
websocket-client==1.8.0
wsproto==1.2.0
XlsxWriter==3.2.0
